# docs/openapi-ui-draft.yaml
openapi: 3.0.3
info:
  title: Smart Vision API (UI Draft)
  version: 0.1.0
  description: |
    OpenAPI-style draft for frontend integration.
    This file is manually maintained for UI generation and may be refined later.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /api/v1/vision/search:
    post:
      tags: [Vision]
      summary: Text search
      operationId: visionSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchQueryDTO'
            examples:
              default:
                value:
                  keyword: "猫咪在沙发上"
                  topK: 30
                  limit: 20
                  searchType: "0"
      responses:
        '200':
          description: Success
          headers:
            X-Strategy-Requested:
              schema: { type: string }
              description: Optional debug header (when app.debug.strategy-header-enabled=true)
            X-Strategy-Effective:
              schema: { type: string }
              description: Optional debug header
            X-Strategy-Fallback:
              schema: { type: string }
              description: Optional debug header
            X-Strategy-Fallback-Reason:
              schema: { type: string }
              description: Optional debug header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResultEnvelope'
  /api/v1/vision/similar:
    get:
      tags: [Vision]
      summary: Similar search by image document id
      operationId: visionSimilar
      parameters:
        - in: query
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResultEnvelope'
...
  /api/v1/vision/search-by-image:
    post:
      tags: [Vision]
      summary: Search by uploaded image
      operationId: visionSearchByImage
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Success
          headers:
            X-Strategy-Requested:
              schema: { type: string }
            X-Strategy-Effective:
              schema: { type: string }
            X-Strategy-Fallback:
              schema: { type: string }
            X-Strategy-Fallback-Reason:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResultEnvelope'
  /api/v1/vision/analyze/stream:
    post:
      tags: [Vision]
      summary: Stream image analysis result
      operationId: visionAnalyzeByImageStream
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                mode:
                  type: string
                  default: general
                enableOcr:
                  type: boolean
                  default: true
                enableGraph:
                  type: boolean
                  default: true
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  securitySchemes:
    AdminSecretHeader:
      type: apiKey
      in: header
      name: X-Admin-Secret
    AccessTokenHeader:
      type: apiKey
      in: header
      name: X-Access-Token
  schemas:
    SearchQueryDTO:
      type: object
      properties:
        keyword:
          type: string
          maxLength: 50
          description: Required by backend controller logic
        topK:
          type: integer
        pageNo:
          type: integer
        similarity:
          type: number
          format: float
          deprecated: true
        enableOcr:
          type: boolean
        searchType:
          type: string
          description: "0:HYBRID, 1:VECTOR_ONLY, 2:TEXT_ONLY, 3:IMAGE_TO_IMAGE"
          enum: ["0", "1", "2", "3"]
        limit:
          type: integer
        searchAfter:
          type: array
          items: {}
    SearchResultDTO:
      type: object
      properties:
        id: { type: string }
        url: { type: string }
        score:
          type: number
          format: double
        ocrText: { type: string }
        highlight: { type: string }
        filename: { type: string }
        sortValues:
          type: array
          items: {}
        tags:
          type: array
          items: { type: string }
...
    SearchResultEnvelope:
      type: object
      properties:
        code: { type: integer, example: 200 }
        message: { type: string, example: Success }
        data:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultDTO'
        timestamp: { type: integer, format: int64 }
...